requirements.txt
asgiref==3.7.2
attrs==23.1.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.4
django-filter==23.5
django-import-export==3.3.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.26.5
drf-yasg==1.21.7
et-xmlfile==1.1.0
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.31.1
rpds-py==0.13.2
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.7.1
uritemplate==4.1.1
xlrd==2.0.1
xlwt==1.3.0
